change db structure

+++ MoneyCounterSite/models.py
@@ -48,7 +48,7 @@ class Party(models.Model):
     datetime = models.DateTimeField(verbose_name = 'Дата проведения', db_index = True, blank = True, null = True)
     place = models.CharField(max_length=200, verbose_name = 'Место проведения')
     persons = models.ManyToManyField(to=Profile, verbose_name = 'Участники')
-    total_cost = models.FloatField(verbose_name = 'Общая стоимость тусы', db_index = True, blank = True, null = True)
+    total_cost = models.DecimalField(max_digits=5, decimal_places=2, verbose_name = 'Общая стоимость тусы', db_index = True, blank = True, null = True)
     #likes = GenericRelation('MoneyCounterSite.Like', related_query_name='parties',  blank=True, null=True)
 
     def __str__(self):
@@ -67,7 +67,7 @@ class Payment(models.Model):
     user = models.ForeignKey(to=Profile, verbose_name = 'Чей платеж')
     party = models.ForeignKey(to=Party, verbose_name = 'Для какой тусовки')
     description = models.CharField(max_length=200, verbose_name = 'Описание платежа')
-    cost = models.FloatField(verbose_name = 'Сколько заплатил')
+    cost = models.DecimalField(max_digits=5, decimal_places=2, verbose_name = 'Сколько заплатил')
 
 
     class Meta:
